moved teamviewset from core to curriculum tracking + tests
