+++ backend/core/views.py
@@ -6,22 +6,16 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAdminUser, IsAuthenticated
 from rest_framework.response import Response
 from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework import permissions as drf_permissions
 from core import permissions as core_permissions
-from core.filters import ObjectPermissionsFilter
 from core.models import Team
-from rest_framework import viewsets, status
-from core.permissions import HasObjectPermission
-from curriculum_tracking.serializers import (
-    TeamStatsSerializer,
-    CardSummarySerializer,
-    UserDetailedStatsSerializer,
-)
+from rest_framework import viewsets
+from curriculum_tracking.serializers import UserDetailedStatsSerializer
 
 # TODO: REFACTOR. If the management helper is used ourtside the management dir then it should be moved
 from curriculum_tracking.management.helpers import get_team_cards
 from django.contrib.postgres.aggregates import StringAgg
 from django.utils import timezone
+
 
 
 @api_view(["POST"])
@@ -90,93 +84,6 @@ class UserProfileViewSet(viewsets.ModelViewSet):
 class CurriculumViewSet(viewsets.ModelViewSet):
     queryset = models.Curriculum.objects.all().order_by("name")
     serializer_class = serializers.CurriculumSerializer
-
-
-class TeamViewSet(viewsets.ModelViewSet):
-    serializer_class = serializers.TeamSerializer
-    filter_backends = [
-        DjangoFilterBackend,
-        ObjectPermissionsFilter(models.Team.PERMISSION_VIEW),
-    ]
-    filterset_fields = ["active"]
-
-    permission_classes = [
-        drf_permissions.IsAuthenticated and core_permissions.IsReadOnly
-    ]
-
-    def get_queryset(self):
-        queryset = (
-            models.Team.objects.all()
-            .order_by("name")
-            .prefetch_related("user_set")
-            # .prefetch_related("team_memberships__user")
-        )
-        return queryset
-
-    @action(
-        detail=False,
-        methods=["GET"],
-        serializer_class=TeamStatsSerializer,
-        permission_classes=[
-            IsAdminUser
-            | core_permissions.HasObjectPermission(
-                permissions=models.Team.PERMISSION_VIEW,
-            )
-        ],
-    )
-    def summary_stats(self, request, pk=None):
-
-        page = self.paginate_queryset(self.filter_queryset(self.get_queryset()))
-        if page is not None:
-            serializer = self.get_serializer(page, many=True)
-            return self.get_paginated_response(serializer.data)
-
-        serializer = self.get_serializer(self.get_queryset(), many=True)
-        return Response(serializer.data)
-
-        # serializer = self.get_serializer(data=request.data)
-        # if serializer.is_valid():
-        #     team = self.get_object()
-        #     return Response(TeamStatsSerializer(team).data)
-        # else:
-        #     return Response(serializer.errors, status="BAD_REQUEST")
-
-    # @action(
-    #     detail=True,
-    #     methods=["get"],
-    #     permission_classes=[
-    #         core_permissions.HasObjectPermission(
-    #             models.Team.PERMISSION_ASSIGN_REVIEWERS
-    #         )
-    #     ],
-    # )
-    # def shuffle_reviewers(self, request, pk=None):
-    #     return Response("TODO")
-
-    @action(
-        detail=True,
-        methods=["POST"],
-        serializer_class=serializers.BulkSetDueTimeSerializer,
-        permission_classes=[
-            HasObjectPermission(permissions=Team.PERMISSION_MANAGE_CARDS)
-        ],
-    )
-    def bulk_set_due_dates(self, request, pk=None):
-
-        team: models.Team = self.get_object()
-        serializer = self.get_serializer(data=request.data)
-
-        if serializer.is_valid():
-
-            team_cards = get_team_cards(
-                team, serializer.validated_data.get("content_item")
-            )
-            for card in team_cards:
-                if card.flavour_ids_match(serializer.validated_data.get("flavours")):
-
-                    card.set_due_time(request.data.get("due_time"))
-
-        return Response([CardSummarySerializer(card).data for card in team_cards])
 
 
 def _get_teams_from_user(self, request, view):
