+++ backend/curriculum_tracking/api_views.py
@@ -28,6 +28,14 @@ from django.db.models import Q, Max
 from django.db.models import Count
 from sql_util.utils import SubqueryAggregate
 from core import serializers as core_serializers
+from core.filters import ObjectPermissionsFilter
+from rest_framework import permissions as drf_permissions
+from curriculum_tracking.serializers import (
+    TeamStatsSerializer,
+    CardSummarySerializer,
+)
+from rest_framework.permissions import IsAdminUser
+from .management.helpers import get_team_cards
 
 
 def _get_teams_from_topic_progress(self, request, view):
@@ -1226,3 +1234,90 @@ class CurriculumContentRequirementViewset(viewsets.ModelViewSet):
 # from django.contrib.postgres.aggregates import *
 # # class UserReviewPerformance(viewsets.ModelViewSet):
 #     RecruitProjectReview.objects.filter(reviewer_user__email="vuyisanani.meteni@umuzi.org").filter(timestamp__gte = timezone.now() - timezone.timedelta(days=7)).annotate(flavour_names=StringAgg('recruit_project__flavours__name',delimiter=",", ordering= 'recruit_project__flavours__name')).values('id','flavour_names','recruit_project__content_item_id')
+
+
+class TeamViewSet(viewsets.ModelViewSet):
+    serializer_class = core_serializers.TeamSerializer
+    filter_backends = [
+        DjangoFilterBackend,
+        ObjectPermissionsFilter(models.Team.PERMISSION_VIEW),
+    ]
+    filterset_fields = ["active"]
+
+    permission_classes = [
+        drf_permissions.IsAuthenticated and core_permissions.IsReadOnly
+    ]
+
+    def get_queryset(self):
+        queryset = (
+            models.Team.objects.all()
+            .order_by("name")
+            .prefetch_related("user_set")
+            # .prefetch_related("team_memberships__user")
+        )
+        return queryset
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        serializer_class=TeamStatsSerializer,
+        permission_classes=[
+            IsAdminUser
+            | core_permissions.HasObjectPermission(
+                permissions=models.Team.PERMISSION_VIEW,
+            )
+        ],
+    )
+    def summary_stats(self, request, pk=None):
+
+        page = self.paginate_queryset(self.filter_queryset(self.get_queryset()))
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+            return self.get_paginated_response(serializer.data)
+
+        serializer = self.get_serializer(self.get_queryset(), many=True)
+        return Response(serializer.data)
+
+        # serializer = self.get_serializer(data=request.data)
+        # if serializer.is_valid():
+        #     team = self.get_object()
+        #     return Response(TeamStatsSerializer(team).data)
+        # else:
+        #     return Response(serializer.errors, status="BAD_REQUEST")
+
+    # @action(
+    #     detail=True,
+    #     methods=["get"],
+    #     permission_classes=[
+    #         core_permissions.HasObjectPermission(
+    #             models.Team.PERMISSION_ASSIGN_REVIEWERS
+    #         )
+    #     ],
+    # )
+    # def shuffle_reviewers(self, request, pk=None):
+    #     return Response("TODO")
+
+    @action(
+        detail=True,
+        methods=["POST"],
+        serializer_class=core_serializers.BulkSetDueTimeSerializer,
+        permission_classes=[
+            HasObjectPermission(permissions=Team.PERMISSION_MANAGE_CARDS)
+        ],
+    )
+    def bulk_set_due_dates(self, request, pk=None):
+
+        team: models.Team = self.get_object()
+        serializer = self.get_serializer(data=request.data)
+
+        if serializer.is_valid():
+
+            team_cards = get_team_cards(
+                team, serializer.validated_data.get("content_item")
+            )
+            for card in team_cards:
+                if card.flavour_ids_match(serializer.validated_data.get("flavours")):
+
+                    card.set_due_time(request.data.get("due_time"))
+
+        return Response([CardSummarySerializer(card).data for card in team_cards])
