+++ backend/curriculum_tracking/tests/test_RecruitProject.py
@@ -1,4 +1,5 @@
-from core.tests.factories import TeamFactory, UserFactory
+from core.tests.factories import UserFactory
+from curriculum_tracking.tests.factories import TeamFactory
 from django.test import TestCase
 
 from curriculum_tracking import models
