+++ backend/curriculum_tracking/tests/test_api_views.py
@@ -4,7 +4,7 @@ from git_real.tests.factories import PullRequestFactory, PullRequestReviewFactor
 from rest_framework.test import APITestCase
 from guardian.shortcuts import assign_perm
 from test_mixins import APITestCaseMixin
-from core.tests.factories import UserFactory, TeamFactory
+from core.tests.factories import UserFactory
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.dateparse import parse_datetime
@@ -28,6 +28,7 @@ from curriculum_tracking.tests.factories import (
 )
 from curriculum_tracking.management.helpers import get_team_cards
 from core.models import Team
+from django.contrib.auth.models import Group
 
 
 class CardSummaryViewsetTests(APITestCase, APITestCaseMixin):
@@ -661,7 +662,7 @@ class BurndownSnapShotViewsetTests(APITestCase, APITestCaseMixin):
 
     def setUp(self):
         self.api_url = self.get_list_url()
-        self.team1 = TeamFactory()
+        self.team1 = factories.TeamFactory()
         self.team1_users = [UserFactory() for _ in range(3)]
 
     def test_staff_member_can_view_all_burndown_snapshot_objects(self):
@@ -692,7 +693,7 @@ class BurndownSnapShotViewsetTests(APITestCase, APITestCaseMixin):
             self.assertEqual(response.status_code, 200)
 
     def test_team1_users_cannot_view_team2_burndown_snapshot_objects(self):
-        team2 = TeamFactory()
+        team2 = factories.TeamFactory()
         team2_users = [UserFactory() for _ in range(3)]
 
         for user in self.team1_users:
@@ -709,8 +710,8 @@ class TestBulkSetDueDatesApi(APITestCase, APITestCaseMixin):
     SUPPRESS_TEST_GET_LIST = True
 
     def setUp(self):
-        self.blue_team = core_factories.TeamFactory(name="BLUE TEAM")
-        self.red_team = core_factories.TeamFactory(name="RED TEAM")
+        self.blue_team = factories.TeamFactory(name="BLUE TEAM")
+        self.red_team = factories.TeamFactory(name="RED TEAM")
         self.user_one_blue = core_factories.UserFactory(
             first_name="one_blue", is_superuser=False, is_staff=False
         )
@@ -997,7 +998,7 @@ class TestCompetenceReviewQueueViewSet(APITestCase, APITestCaseMixin):
             project.reviewer_users.add(UserFactory())
             project.request_review()
 
-        teams = [TeamFactory() for _ in range(5)]
+        teams = [factories.TeamFactory() for _ in range(5)]
 
         teams[0].user_set.add(cards[0].recruit_project.recruit_users.first())
         teams[1].user_set.add(cards[1].recruit_project.recruit_users.first())
@@ -1052,3 +1053,72 @@ class CurriculumContentRequirementViewsetTests(APITestCase, APITestCaseMixin):
 
     def verbose_instance_factory(self):
         return factories.CurriculumContentRequirementFactory()
+
+
+class TestTeamViewSet(APITestCase, APITestCaseMixin):
+    LIST_URL_NAME = "team-list"
+    SUPPRESS_TEST_POST_TO_CREATE = True
+
+    def verbose_instance_factory(self):
+        team = factories.TeamFactory()
+        user = factories.UserFactory()
+        team.user_set.add(user)
+        return team
+
+    def test_only_returns_teams_i_can_see(self):
+
+        url = self.get_list_url()
+        super_user = factories.UserFactory(is_superuser=True)
+        staff_user = factories.UserFactory(is_staff=True)
+        normal_user = factories.UserFactory()
+
+        team_1 = factories.TeamFactory()
+        factories.TeamFactory()
+
+        self.login(super_user)
+
+        response = self.client.get(url)
+        self.assertEqual(len(response.data), 2)
+
+        self.login(staff_user)
+
+        response = self.client.get(url)
+        self.assertEqual(len(response.data), 0)
+
+        for permission, _ in Team._meta.permissions:
+            # make sure that users can read what they need to
+            user = factories.UserFactory()
+            assign_perm(permission, user, team_1)
+            self.login(user)
+
+            response = self.client.get(url)
+            self.assertEqual(len(response.data), 1)
+            self.assertEqual(response.data[0]["id"], team_1.id)
+
+            # check that it works for groups as well
+            group = Group.objects.create(name=f"{permission} group")
+            user = factories.UserFactory()
+            group.user_set.add(user)
+            assign_perm(permission, group, team_1)
+
+            self.login(user)
+
+            response = self.client.get(url)
+            self.assertEqual(len(response.data), 1)
+            self.assertEqual(response.data[0]["id"], team_1.id)
+
+        self.login(normal_user)
+
+        response = self.client.get(url)
+        self.assertEqual(len(response.data), 0)
+
+    def test_staff_users_cant_see_all_teams(self):
+        user = factories.UserFactory(is_superuser=False, is_staff=True)
+
+        factories.TeamFactory()
+
+        self.login(user)
+        url = self.get_list_url()
+        response = self.client.get(url)
+        self.assertEqual(len(response.data), 0)
+
